# docs/api/openapi-v1.yml
openapi: 3.0.3
info:
  title: TaskCare REST API
  version: v1
  description: |
    This is the RESTful API of the TaskCare backend service. TaskCare allows you to
    manage and organize tasks collaboratively and in Kanban style. Via this interface it is possible
    to call up data from the task management platform synchronously. This includes
    data, such as users and their roles, as well as content data like boards and tasks.
    
    # Overview
    
    The TaskCare API is a machine interface which is RESTful over HTTP. The
    following specifications apply API-wide and to all endpoints of the interface.
    
    ## HTTP Verbs
    
    There are a couple of possible HTTP verbs used for the Content API. The following
    table clearly clarifies their use and meaning within the API.
    
    | Verb         | Description                                                       |
    |--------------|-------------------------------------------------------------------|
    | HEAD         | Can be ran against any resource to get just the HTTP header info. |
    | GET          | Used for retrieving resources.                                    |
    | POST         | Used for creating resources.                                      |
    | PATCH        | Used for updating resources partially.                            |
    | PUT          | Used for replacing resources or collections.                      |
...
    
    ## Content Negotiation
    
...
    
    ## Timestamps
    
...
    
...

...

...

...
servers:
...
paths:
...
  /boards/{boardId}/statuses:
    get:
      tags:
        - Status
      summary: Returns all statuses of a board.
      description: >
        Returns all statuses of a board.
      parameters:
        - name: boardId
          in: path
          description: Unique identifier of the board.
          required: true
          schema:
            type: string
            format: uuid
            example: 3377cf8f-da29-4aa7-9036-91951b8fce87
        - name: page
          in: query
          description: Zero based index of page to load.
          required: false
          schema:
            type: integer
            example: 0
        - name: perPage
          in: query
          description: Number of elements per page.
          required: false
          schema:
            type: integer
            example: 10
        - name: search
          in: query
          description: Search query for filtering collection.
          required: false
          schema:
            type: string
            format: rsql
      responses:
        200:
          description: Successfully fetched all members of the board.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberPageDto"
              example:
                page: 0
                perPage: 10
                totalElements: 1
                totalPages: 1
                content:
                  - id: 037f610b-f09c-4648-8c08-39a3ade3f447
                    name: To Do
                    description: >
                      Tasks that need to be done.
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
    post:
      tags:
        - Status
      summary: Creates a new status.
      description: >
        Creates a new status. Only available for board administrators.
      parameters:
        - name: boardId
          in: path
          description: Unique identifier of the board.
          required: true
          schema:
            type: string
            format: uuid
            example: 3377cf8f-da29-4aa7-9036-91951b8fce87
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddStatusDto"
            example:
              name: To Do
              description: >
                Tasks that need to be done.
      responses:
        201:
          description: Successfully created the status.
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
  /boards/{boardId}/statuses/{statusId}:
    get:
      tags:
        - Status
      summary: Returns a single status.
      description: >
        Returns a single status.
      parameters:
        - name: boardId
          in: path
          description: Unique identifier of the board.
          required: true
          schema:
            type: string
            format: uuid
            example: 3377cf8f-da29-4aa7-9036-91951b8fce87
        - name: statusId
          in: path
          description: Unique identifier of the status.
          required: true
          schema:
            type: string
            format: uuid
            example: 037f610b-f09c-4648-8c08-39a3ade3f447
      responses:
        200:
          description: Successfully fetched the status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusDto"
              example:
                id: 037f610b-f09c-4648-8c08-39a3ade3f447
                name: To Do
                description: >
                  Tasks that need to be done.
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
    delete:
      tags:
        - Status
      summary: Deletes a single status.
      description: >
        Deletes a single status. Only available for board administrators.
      parameters:
        - name: boardId
          in: path
          description: Unique identifier of the board.
          required: true
          schema:
            type: string
            format: uuid
            example: 3377cf8f-da29-4aa7-9036-91951b8fce87
        - name: statusId
          in: path
          description: Unique identifier of the status.
          required: true
          schema:
            type: string
            format: uuid
            example: 037f610b-f09c-4648-8c08-39a3ade3f447
      responses:
        204:
          description: Successfully deleted the status.
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
    patch:
      tags:
        - Status
      summary: Updates the details of a single status.
      description: >
        Updates a status. Only available for board administrators.
      parameters:
        - name: boardId
          in: path
          description: Unique identifier of the board.
          required: true
          schema:
            type: string
            format: uuid
            example: 3377cf8f-da29-4aa7-9036-91951b8fce87
        - name: statusId
          in: path
          description: Unique identifier of the status.
          required: true
          schema:
            type: string
            format: uuid
            example: 037f610b-f09c-4648-8c08-39a3ade3f447
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusDto"
            example:
              name: To Do
              description: >
                Tasks that need to be done.
      responses:
        200:
          description: Successfully updated the status.
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
components:
  schemas:
    ValidationErrorDetailsDto:
      type: object
      properties:
        field:
          type: string
          description: The name of the field.
        message:
          type: string
          description: The error message.
        code:
          type: string
          description: The validation error code.
      required:
        - field
        - message
        - code
    InvalidParameterErrorDetailsDto:
      type: object
      properties:
        parameter:
          type: string
          description: The name of the parameter.
        message:
          type: string
          description: The error message.
      required:
        - parameter
        - message
    ErrorDto:
      type: object
      properties:
        error:
          type: string
          description: The type of error.
        status:
          type: integer
          description: The HTTP status code number.
        timestamp:
          type: string
          format: date-time
          description: The time the error occurred.
        path:
          type: string
          description: The endpoint path on which the error occurred.
        details:
          description: Further information on the error or its cause.
          type: array
          items:
            oneOf:
              - $ref: "#/components/schemas/ValidationErrorDetailsDto"
              - $ref: "#/components/schemas/InvalidParameterErrorDetailsDto"
      required:
        - error
        - status
        - timestamp
        - path
    PageDto:
      type: object
      properties:
        page:
          type: integer
          description: Zero-based index of the current page.
        perPage:
          type: integer
          description: Number of elements per page.
        totalElements:
          type: integer
          description: Number of all elements of the collection.
        totalPages:
          type: integer
          description: Number of all pages into which the collection was divided.
      required:
        - page
        - perPage
        - totalElements
        - totalPages
...
    MemberPageDto:
      allOf:
        - $ref: "#/components/schemas/PageDto"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/MemberDto"
          required:
            - content
    MemberDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of member.
        userId:
          type: string
          format: uuid
          description: Unique identifier of user.
        role:
          type: string
          description: Role of user.
          enum:
            - MEMBER
            - MAINTAINER
            - ADMINISTRATOR
      required:
        - id
        - userId
        - role
...
    StatusDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of member.
        name:
          type: string
          description: Name of status.
        description:
          type: string
          description: Description of status.
      required:
        - id
        - name
...
    AddStatusDto:
      type: object
      properties:
        name:
          type: string
          description: Name of status.
        description:
          type: string
          description: Description of status.
      required:
        - name
    UpdateStatusDto:
      type: object
      properties:
        name:
          type: string
          description: Name of status.
        description:
          type: string
          description: Description of status.
  responses:
...
    403:
      description: Access was denied due to insufficient rights.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDto"
    401:
      description: The access must be authenticated or authentication failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDto"
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: jwt
security:
  - jwt: [ ]